Move heuristic gesture HUD to samples
