+++ samples/gestures_heuristic/README.md
@@ -1,0 +1,9 @@
+# Heuristic Gesture HUD
+
+This sample enables the heuristic gesture recognizer and shows a simple HUD that
+displays the most confident gesture per hand while logging gesture start/end
+events to the console. Use it to validate pinch, open-palm, fist, thumbs-up,
+point, and spread detections in the simulator or on device.
+
+Run `npm run serve` (or your preferred static server) and open
+`samples/gestures_heuristic/` in a WebXR-capable browser.
