+++ templates/heuristic_hand_gestures/README.md
@@ -6,3 +6,5 @@ open-palm, fist, thumbs-up, point, and spread detections on devices like Quest
 or in the simulator.
 
 For gesture configuration details, see https://xrblocks.github.io/docs/manual/HandGestures/.
+
+For an on-screen HUD and richer feedback, see `samples/gestures_heuristic/`.
